Move CLI logic out of cmd/cli/main.go in order to keep main.go simple
I initially inteded to have cli.go live in cmd/cli but this meant compiling should be done as `cd cmd/cli && go build .`. The binary would also then be created in cmd/cli, which is not what I want. I would prefer to have it in bin/cli. This could of course be achieved with a simple `mv`, but this makes the build command unecessarily complicated, so I chose to go with this structure instead, for now.